+++ crates/miden-testing/src/kernel_tests/block/utils.rs
@@ -9,7 +9,9 @@ use miden_objects::{
     batch::ProvenBatch,
     block::BlockNumber,
     note::{Note, NoteId, NoteTag, NoteType},
-    testing::{account_component::AccountMockComponent, note::NoteBuilder},
+    testing::{
+        account_component::AccountMockComponent, account_id::ACCOUNT_ID_SENDER, note::NoteBuilder,
+    },
     transaction::{ExecutedTransaction, OutputNote, ProvenTransaction, TransactionScript},
     utils::word_to_masm_push_string,
 };
@@ -63,7 +65,7 @@ pub fn generate_output_note(sender: AccountId, seed: [u8; 32]) -> Note {
     NoteBuilder::new(sender, &mut rng)
         .note_type(NoteType::Private)
         .tag(NoteTag::for_local_use_case(0, 0).unwrap().into())
-        .build(&TransactionKernel::assembler())
+        .build(&TransactionKernel::assembler().with_debug_mode(true))
         .unwrap()
 }
 
@@ -143,10 +145,30 @@ pub fn generate_tx_with_authenticated_notes(
     ProvenTransaction::from_executed_transaction_mocked(executed_tx)
 }
 
+/// Generates a NOOP transaction, i.e. one that doesn't change the state of the account.
+///
+/// To make this transaction non-empty, it consumes one "noop note", which does nothing.
+pub fn generate_noop_tx(
+    chain: &mut MockChain,
+    input: impl Into<TxContextInput>,
+) -> ExecutedTransaction {
+    let noop_note = NoteBuilder::new(ACCOUNT_ID_SENDER.try_into().unwrap(), &mut rand::rng())
+        .build(&TransactionKernel::assembler())
+        .expect("failed to create the noop note");
+    chain.add_pending_note(OutputNote::Full(noop_note.clone()));
+    chain.prove_next_block();
+
+    let tx_context = chain
+        .build_tx_context(input.into(), &[noop_note.id()], &[])
+        .input_notes(vec![noop_note])
+        .build();
+    tx_context.execute().unwrap()
+}
+
 /// Generates a transaction that expires at the given block number.
 pub fn generate_tx_with_expiration(
     chain: &mut MockChain,
-    account_id: AccountId,
+    input: impl Into<TxContextInput>,
     expiration_block: BlockNumber,
 ) -> ProvenTransaction {
     let expiration_delta = expiration_block
@@ -154,7 +176,7 @@ pub fn generate_tx_with_expiration(
         .unwrap();
 
     let tx_context = chain
-        .build_tx_context(account_id, &[], &[])
+        .build_tx_context(input, &[], &[])
         .tx_script(authenticate_mock_account_tx_script(expiration_delta.as_u32() as u16))
         .build();
     let executed_tx = tx_context.execute().unwrap();