+++ crates/miden-tx/src/executor/notes_checker.rs
@@ -19,7 +19,7 @@ use super::{NoteAccountExecution, TransactionExecutor, TransactionExecutorError}
 /// The check is performed using the [NoteConsumptionChecker::check_notes_consumability] procedure.
 /// Essentially runs the transaction to make sure that provided input notes could be consumed by the
 /// account.
-pub struct NoteConsumptionChecker<'a>(&'a TransactionExecutor);
+pub struct NoteConsumptionChecker<'a>(&'a TransactionExecutor<'a, 'a>);
 
 impl<'a> NoteConsumptionChecker<'a> {
     /// Creates a new [`NoteConsumptionChecker`] instance with the given transaction executor.
@@ -89,7 +89,7 @@ impl<'a> NoteConsumptionChecker<'a> {
 
         // if all checked notes turned out to be either `P2ID` or `P2IDR` notes and all of them
         // passed, then we could safely return the `Success`
-        if successful_notes.len() == input_notes.num_notes() {
+        if successful_notes.len() == (input_notes.num_notes() as usize) {
             return Ok(NoteAccountExecution::Success);
         }